+++ hackathon_site/event/api_urls.py
@@ -30,11 +30,6 @@ urlpatterns = [
         name="incident-list",
     ),
     path(
-        "teams/team/incidents/<int:pk>/",
-        api_views.TeamIncidentDetailView.as_view(),
-        name="incident-detail",
-    ),
-    path(
         "teams/team/orders/",
         api_views.CurrentTeamOrderListView.as_view(),
         name="team-orders",
