+++ hackathon_site/event/test_api.py
@@ -693,90 +693,6 @@ class TeamIncidentListViewPostTestCase(SetupUserMixin, APITestCase):
             self.assertEqual(final_response[attribute], self.request_data[attribute])
 
 
-class TeamIncidentDetailViewGetTestCase(SetupUserMixin, APITestCase):
-    def setUp(self):
-        super().setUp()
-        self.group = Group.objects.create(name="Test Users")
-        self.user.groups.add(self.group)
-        self.profile = Profile.objects.create(user=self.user)
-        self.team2 = Team.objects.create()
-        self.team = Team.objects.create()
-        self.order = Order.objects.create(
-            status="Cart",
-            team=self.team,
-            request={"hardware": [{"id": 1, "quantity": 2}, {"id": 2, "quantity": 3}]},
-        )
-        self.hardware = Hardware.objects.create(
-            name="name",
-            model_number="model",
-            manufacturer="manufacturer",
-            datasheet="/datasheet/location/",
-            notes="notes",
-            quantity_available=4,
-            max_per_team=1,
-            picture="/picture/location",
-        )
-
-        self.other_hardware = Hardware.objects.create(
-            name="other",
-            model_number="otherModel",
-            manufacturer="otherManufacturer",
-            datasheet="/datasheet/location/other",
-            quantity_available=10,
-            max_per_team=10,
-            picture="/picture/location/other",
-        )
-        self.order_item = OrderItem.objects.create(
-            order=self.order, hardware=self.hardware,
-        )
-
-        self.order_item2 = OrderItem.objects.create(
-            order=self.order, hardware=self.other_hardware,
-        )
-
-        self.incident = Incident.objects.create(
-            state="Broken",
-            description="Description",
-            order_item=self.order_item,
-            time_occurred=datetime(2022, 8, 8, tzinfo=settings.TZ_INFO),
-        )
-
-        self.another_incident = Incident.objects.create(
-            state="Missing",
-            description="Description",
-            order_item=self.order_item2,
-            time_occurred=datetime(2022, 1, 1, tzinfo=settings.TZ_INFO),
-        )
-
-        self.view = reverse("api:event:incident-detail", args=[self.incident.id])
-
-    def test_user_not_logged_in(self):
-        response = self.client.get(self.view)
-        self.assertEqual(response.status_code, status.HTTP_401_UNAUTHORIZED)
-
-    def test_user_has_no_profile(self):
-        self.profile.delete()
-        self._login()
-        response = self.client.get(self.view)
-        self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
-
-    def test_user_has_profile(self):
-        self._login()
-        response = self.client.get(self.view)
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-
-    def test_incident_get_success(self):
-        self._login()
-        response = self.client.get(self.view)
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-
-        queryset = Incident.objects.get(id=self.incident.id)
-        # need to provide a request in the serializer context to produce absolute url for image field
-        expected_response = IncidentSerializer(queryset).data
-        data = response.json()
-        self.assertEqual(expected_response, data)
-
-
 class EventTeamDetailViewTestCase(SetupUserMixin, APITestCase):
     def setUp(self, **kwargs):
         self.team = Team.objects.create()
